# Add transaction to batch insert