--- a/pkg/dotc1z/sql_helpers.go
+++ b/pkg/dotc1z/sql_helpers.go
@@ -356,7 +356,7 @@ func executeChunkedInsert(
 
 	if txError != nil {
 		if rollbackErr := tx.Rollback(); rollbackErr != nil {
-			return fmt.Errorf("error rolling back transaction after error: %w, original error: %v", rollbackErr, txError)
+			return fmt.Errorf("error rolling back transaction after error: %w, original error: %w", rollbackErr, txError)
 		}
 
 		return fmt.Errorf("error executing chunked insert: %w", txError)