--- a/pkg/dotc1z/sql_helpers_test.go
+++ b/pkg/dotc1z/sql_helpers_test.go
@@ -1,0 +1,102 @@
+package dotc1z
+
+import (
+	"context"
+	v2 "github.com/conductorone/baton-sdk/pb/c1/connector/v2"
+	"github.com/conductorone/baton-sdk/pkg/types/resource"
+	"github.com/segmentio/ksuid"
+	"github.com/stretchr/testify/require"
+	"path/filepath"
+	"testing"
+)
+
+func generateResources(count int, resourceType *v2.ResourceType) []*v2.Resource {
+	if count < 1 {
+		return nil
+	}
+
+	response := make([]*v2.Resource, count)
+
+	for i := range count {
+		id := ksuid.New().String()
+
+		newResource, err := resource.NewResource(id, resourceType, id)
+		if err != nil {
+			panic(err)
+		}
+
+		response[i] = newResource
+	}
+
+	return response
+}
+
+func TestPutResources(t *testing.T) {
+	ctx := context.Background()
+
+	tempDir := filepath.Join(t.TempDir(), "test.c1z")
+
+	c1zFile, err := NewC1ZFile(ctx, tempDir, WithPragma("journal_mode", "WAL"))
+	require.NoError(t, err)
+
+	_, err = c1zFile.StartNewSync(ctx)
+	require.NoError(t, err)
+
+	err = c1zFile.PutResources(
+		ctx,
+		generateResources(10_000, &v2.ResourceType{
+			Id:          "test_resource_type",
+			DisplayName: "Test Resource Type",
+		})...,
+	)
+	require.NoError(t, err)
+
+	err = c1zFile.Close()
+	require.NoError(t, err)
+}
+
+func BenchmarkPutResources(b *testing.B) {
+	cases := []struct {
+		name           string
+		resourcesCount int
+	}{
+		{"500", 500},
+		{"1k", 1_000},
+		{"10k", 10_000},
+		{"50k", 50_000},
+	}
+
+	for _, c := range cases {
+		b.Run(c.name, func(b *testing.B) {
+			b.ReportAllocs()
+			for i := 0; i < b.N; i++ {
+				b.StopTimer()
+				ctx := context.Background()
+
+				tempDir := filepath.Join(b.TempDir(), "test.c1z")
+
+				c1zFile, err := NewC1ZFile(ctx, tempDir, WithPragma("journal_mode", "WAL"))
+				require.NoError(b, err)
+
+				_, err = c1zFile.StartNewSync(ctx)
+				require.NoError(b, err)
+
+				generatedData := generateResources(c.resourcesCount, &v2.ResourceType{
+					Id:          "test_resource_type",
+					DisplayName: "Test Resource Type",
+				})
+
+				b.StartTimer()
+
+				err = c1zFile.PutResources(
+					ctx,
+					generatedData...,
+				)
+				require.NoError(b, err)
+
+				err = c1zFile.Close()
+				require.NoError(b, err)
+			}
+		})
+	}
+}