Add a transaction to execute the batch

Before
```
goos: darwin
goarch: arm64
pkg: github.com/conductorone/baton-sdk/pkg/dotc1z
cpu: Apple M1 Max
BenchmarkName
BenchmarkName/500
BenchmarkName/500-10         	     159	   7944809 ns/op	20500388 B/op	   26861 allocs/op
BenchmarkName/1k
BenchmarkName/1k-10          	      80	  13110462 ns/op	21841604 B/op	   53592 allocs/op
BenchmarkName/10k
BenchmarkName/10k-10         	       7	 153204155 ns/op	43281107 B/op	  534727 allocs/op
BenchmarkName/50k
BenchmarkName/50k-10         	       1	1117222750 ns/op	138147560 B/op	 2673078 allocs/op
PASS
```

After
```
goos: darwin
goarch: arm64
pkg: github.com/conductorone/baton-sdk/pkg/dotc1z
cpu: Apple M1 Max
BenchmarkName
BenchmarkName/500
BenchmarkName/500-10         	     160	   8204915 ns/op	20502633 B/op	   26908 allocs/op
BenchmarkName/1k
BenchmarkName/1k-10          	      88	  12675455 ns/op	21831980 B/op	   53644 allocs/op
BenchmarkName/10k
BenchmarkName/10k-10         	       9	 119945125 ns/op	43285008 B/op	  534870 allocs/op
BenchmarkName/50k
BenchmarkName/50k-10         	       2	 734946604 ns/op	138147168 B/op	 2673599 allocs/op
PASS
```

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- Refactor
  - Chunked inserts now run inside a single database transaction for atomic, all-or-nothing imports; error handling consolidated so failures trigger a unified rollback and successful runs commit once.

- Tests
  - Added end-to-end tests and benchmarks exercising large-scale resource imports (500–50,000 items), validating correctness, measuring performance, and reporting allocations.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->